1.6

+++ src/app/sample/streetlocation/entity/Declare.java
@@ -3,11 +3,11 @@ package app.sample.streetlocation.entity;
 public class Declare {
 	  public static double UserLat=0;
 	  public static double  UserLon=0;
-	  public static int   start_lat=0;
-	  public static int   start_lon=0;
+	  public static long   start_lat=0;
+	  public static long   start_lon=0;
 	  public static String   start_name="";
-	  public static int   end_lat=0;
-	  public static int   end_lon=0;
+	  public static long   end_lat=0;
+	  public static long   end_lon=0;
 	  public static String   end_name="";
 	  public static boolean   type=false;    //false Æðµã true ÖÕµã
 }
